Put some music stuff into place.
I have a basic music slider on the main menu and made a class and object to control the music.
Mixer is working properly, as is the slider, which currently adjusts music  volume.
Placeholder music is in place. Starts out muted for now, since it's not the actual music.
Implemented first stages of transition between main menu and gameplay with placeholder music.
Sound effects are not tied into the mixer yet.